Update Audio.cpp
Added required typecasts to 'uint32_t' (from 'size_t') for use w. 'min()' function.

+++ src/Audio.cpp
@@ -2901,8 +2901,8 @@ void Audio::processLocalFile() {
 
     availableBytes = 16 * 1024; // set some large value
 
-    availableBytes = min(availableBytes, InBuff.writeSpace());
-    availableBytes = min(availableBytes, audiofile.size() - byteCounter);
+    availableBytes = min(availableBytes, (uint32_t)InBuff.writeSpace());
+    availableBytes = min(availableBytes, (uint32_t)audiofile.size() - byteCounter);
     if(m_contentlength){
         if(m_contentlength > getFilePos()) availableBytes = min(availableBytes, m_contentlength - getFilePos());
     }
